requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# YOLOv5 dependencies
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
yolov5==7.0.13

# Video processing
ffmpeg-python==0.2.0

# Notification services
twilio==8.9.0
pushover-complete==1.1.1
sendgrid==6.10.0

# Async and scheduling
asyncio==3.4.3
schedule==1.2.0
python-crontab==3.0.0

# Configuration and logging
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Database
sqlalchemy==2.0.23
sqlite3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Web interface (optional)
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5